// sls-alm-slmap-api/src/common/enums/workflow.enum.ts
export enum WorkflowType {
  TRADE_AREA_DRAWING = 1,
  TRADE_AREA_EDIT = 2,
  TRADE_AREA_DELETE = 3,
  POTENTIAL_PREPARATORY = 4,
  POTENTIAL_EDIT = 5,
  POTENTIAL_DELETE = 6,
}

export enum WorkflowCompleteStatus {
  WAITING = 'W',
  COMPLETE = 'Y',
  INCOMPLETE = 'N',
  CANCELLED = 'C',
}

export enum WorkflowAction {
  CREATE = 1,
  EDIT = 2,
  DELETE = 3,
  SAVE = 4,
  SEND_APPROVE = 5,
  APPROVE = 6,
  REJECT = 7,
  NOT_APPROVE = 8,
  CANCEL = 9,
  PASS = 10,
  SEND_EDIT = 11,
  SEND_REVIEW = 12,
  SEND_QUOTA = 13,
  CREATE_QUOTA_ROUND = 14,
  REVIEW_QUOTA_END = 15,
}

export enum TradeareaStep {
  DRAFT_PENDING = 101,
  DRAFT_OPERATION_APPROVE = 102,
  DRAFT_GM_APPROVE = 103,
  DRAFT_COMPLETE = 104,
  EDIT_PENDING = 201,
  EDIT_OPERATION_APPROVE = 202,
  EDIT_GM_APPROVE = 203,
  EDIT_COMPLETE = 204,
  DELETE_PENDING = 301,
  DELETE_OPERATION_APPROVE = 302,
  DELETE_GM_APPROVE = 303,
  DELETE_COMPLETE = 304,
}